Added support for multiple Aspect Ratios
Initial resolution will now be dynamically set according
to the aspect ratio of the display. Resolution settings should
also change according to the aspect ratio so they cannot be set
to an undesired or unusable state.

ALSO: there is some extra stuff in here that I dont know how to exclude
from the commit, like the settings file and the iml file

+++ src/com/fisherevans/lrk/states/options/menu_items/MenuItemResolution.java
@@ -11,14 +11,14 @@ import com.fisherevans.lrk.states.options.Menu;
  */
 public class MenuItemResolution extends MenuItem
 {
-    private int _scale = 1;
+    private double _scale = 1;
 
     /**
      * @param text
      * @param description
      * @param scale the new scale to set the display window to
      */
-    public MenuItemResolution(String text, String description, int scale)
+    public MenuItemResolution(String text, String description, double scale)
     {
         super(text, description);
         _scale = scale;
